third_party/WebKit/LayoutTests/inspector/console/console-tainted-globals.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

(function() {
    var _originalFunctionCall = Function.prototype.call;
    var _originalFunctionApply = Function.prototype.apply;

    var _overriddenFunctionCall = function() {
        original();
        console.error("FAIL: Function.prototype.call should not be called!");
        var result = _originalFunctionCall.apply(this, arguments);
        overridden();
        return result;
    }

    var _overriddenFunctionApply = function(thisArg, argsArray) {
        original();
        console.error("FAIL: Function.prototype.apply should not be called!");
        var result = _originalFunctionApply.call(this, thisArg, argsArray);
        overridden();
        return result;
    }

    function original()
    {
        Function.prototype.call = _originalFunctionCall;
        Function.prototype.apply = _originalFunctionApply;
    }

    function overridden()
    {
        Function.prototype.call = _overriddenFunctionCall;
        Function.prototype.apply = _overriddenFunctionApply;
    }

    overridden();
})();

function throwGetter()
{
   console.error("FAIL: Should not be called!");
   throw new Error("FAIL");
}

function testOverriddenArrayPushAndMathMax()
{
    Object.defineProperty(Array.prototype, "push", {
        get: throwGetter
    });
    Object.defineProperty(Math, "max", {
        get: throwGetter
    });
    return [1, 2, 3];
}

function testOverriddenConstructorName()
{
    var obj = {};
    obj.constructor = { name: "foo" };
    return obj;
}

function testThrowConstructorName()
{
    var obj = {};
    Object.defineProperty(obj, "constructor", {
        get: throwGetter
    });
    return obj;
}

function testOverriddenIsFinite()
{
    window.isFinite = throwGetter;
    var out;
    (function() {
        out = arguments;
    })("arg1", "arg2");
    return out;
}

function test()
{
    InspectorTest.runTestSuite([
        function evaluateInConsole(next)
        {
            var expressions = [
                "testOverriddenArrayPushAndMathMax()",
                "testOverriddenConstructorName()",
                "testThrowConstructorName()",
                "testOverriddenIsFinite()"
            ];

            function iterate()
            {
                var expr = expressions.shift();
                if (!expr) {
                    InspectorTest.runAfterPendingDispatches(next);
                    return;
                }
                InspectorTest.evaluateInConsole(expr, iterate);
            }
            iterate();
        },

        function testRuntimeAgentCallFunctionOn(next)
        {
            RuntimeAgent.evaluate("({ a : 1, b : 2 })", step1);

            function step1(error, result, wasThrown)
            {
                function sum()
                {
                    return this.a + this.b;
                }
                RuntimeAgent.callFunctionOn(result.objectId, sum.toString(), step2);
            }

            function step2(error, result, wasThrown)
            {
                InspectorTest.assertEquals(3, result.value);
                next();
            }
        },

        function dumpConsoleMessages(next)
        {
            InspectorTest.dumpConsoleMessages();
            next();
        }
    ]);
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that overriding global methods (like Array.prototype.push, Math.max) will not break the inspector.
</p>

</body>
</html>
